Allow crash handler code to return a jump buffer
This allows us to safely jump from the exception handler code to eg.
release locks, as it appears this will be needed on mac due to how
exception ports work
