+++ exception-handler/src/mac.rs
@@ -1,4 +1,5 @@
 mod ffi;
+pub mod jmp;
 mod signal;
 mod state;
