+++ exception-handler/src/windows.rs
@@ -1,3 +1,4 @@
+pub mod jmp;
 mod state;
 
 use crate::Error;
@@ -26,18 +27,6 @@ pub struct ExceptionHandler {
     inner: std::sync::Arc<state::HandlerInner>,
 }
 
-#[derive(Copy, Clone, PartialEq)]
-pub enum HandleDebugExceptions {
-    Yes,
-    No,
-}
-
-impl From<HandleDebugExceptions> for bool {
-    fn from(hde: HandleDebugExceptions) -> bool {
-        hde == HandleDebugExceptions::Yes
-    }
-}
-
 impl ExceptionHandler {
     /// Attaches an exception handler.
     ///
@@ -47,10 +36,7 @@ impl ExceptionHandler {
     pub fn attach(on_crash: Box<dyn crate::CrashEvent>) -> Result<Self, Error> {
         let inner = {
             let mut handlers = state::HANDLER_STACK.lock();
-            let inner = std::sync::Arc::new(state::HandlerInner::new(
-                HandleDebugExceptions::Yes,
-                on_crash,
-            ));
+            let inner = std::sync::Arc::new(state::HandlerInner::new(on_crash));
             handlers.push(std::sync::Arc::downgrade(&inner));
             inner
         };
